Backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
Django==4.2.21
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
openai==1.91.0
openpyxl==3.1.5
pillow==11.2.1
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.1
requests==2.32.4
rpds-py==0.25.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
